# docs/openapi-v1.yaml
openapi: 3.1.0
info:
  title: YL Cloud Open API
  version: 1.0.0
  description: Stable API Key authenticated contract. Production clients should pin /api/v1.
servers:
  - url: /api/v1
security:
  - bearerApiKey: []
paths:
...
  /files:
    get:
      operationId: listFiles
      parameters:
        - {name: parentId, in: query, schema: {type: integer, format: int64}}
        - {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
        - {name: pageSize, in: query, schema: {type: integer, minimum: 1, maximum: 100}}
      responses:
        '200': {description: Paginated files}
...
components:
  securitySchemes:
    bearerApiKey:
      type: http
      scheme: bearer
      bearerFormat: ylk_*
...